init setup for projects
