+++ services/projects/routes.go
@@ -1,0 +1,7 @@
+package projects
+
+import "github.com/gorilla/mux"
+
+func RegisterRoutes(router *mux.Router, h *Handler) {
+	router.HandleFunc("/projects", h.handleGetProjects).Methods("GET")
+}
