+++ services/projects/store.go
@@ -1,0 +1,19 @@
+package projects
+
+import (
+	"database/sql"
+
+	"github.com/norrico31/it210-core-service-backend/entities"
+)
+
+type Store struct {
+	db *sql.DB
+}
+
+func NewStore(db *sql.DB) *Store {
+	return &Store{db: db}
+}
+
+func (s *Store) GetProjects() ([]*entities.Project, error) {
+	return nil, nil
+}
